refactor(parser): move types.rs into main.rs
